pyproject.toml
[project]
name = "seisbai_tools"
version = "1.0.0"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "msgspec",
    "ulid-py",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["seisbai_tools*"]